+++ packages/isaacscript-common/src/functions/array.ts
@@ -233,6 +233,33 @@ export function emptyArray<T>(array: T[]): void {
 }
 
 /**
+ * Helper function to perform a map and a filter at the same time. Similar to `Array.map`, provide a
+ * function that transforms a value, but return `undefined` if the value should be skipped. (Thus,
+ * this function cannot be used in situations where `undefined` can be a valid array element.)
+ *
+ * This function is useful because a map will always produce an array with the same amount of
+ * elements as the original array.
+ *
+ * This is named `filterMap` after the Rust function:
+ * https://doc.rust-lang.org/std/iter/struct.FilterMap.html
+ */
+export function filterMap<OldT, NewT>(
+  array: OldT[],
+  func: (element: OldT) => NewT | undefined,
+): NewT[] {
+  const newArray: NewT[] = [];
+
+  for (const element of array) {
+    const newElement = func(element);
+    if (newElement !== undefined) {
+      newArray.push(newElement);
+    }
+  }
+
+  return newArray;
+}
+
+/**
  * Helper function to get all possible combinations of the given array. This includes the
  * combination of an empty array.
  *